// frontend/src/features/gateways/models/create-device-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Gateways API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface CreateDeviceDto
 */
export interface CreateDeviceDto {
  /**
   *
   * @type {number}
   * @memberof CreateDeviceDto
   */
  uid: number;
  /**
   *
   * @type {string}
   * @memberof CreateDeviceDto
   */
  vendor: string;
  /**
   *
   * @type {string}
   * @memberof CreateDeviceDto
   */
  status: string;
}